requirements.txt
asgiref==3.6.0
cffi==1.15.1
channels==4.0.0
cryptography==40.0.1
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
sqlparse==0.4.3
